Float32 interface bug

+++ client/tcp.go
@@ -9,8 +9,8 @@ import (
 
 func SetHeaderVal(b []byte, offset int, order binary.ByteOrder, val interface{}) *bytes.Buffer {
 	p := bytes.NewBuffer(b[:offset])
-	log.WithFields(log.Fields{"Receive time": val.(float64)}).Info("time being sent back")
-	err := binary.Write(p, order, val.(float64))
+	log.WithFields(log.Fields{"Receive time": val.(float32)}).Info("time being sent back")
+	err := binary.Write(p, order, val.(float32))
 	if err != nil {
 		log.Fatal("binary.write failed")
 	}
